pulled branch compare components to separate directory
